requirements.txt
aiofiles==23.2.1
amqp==5.2.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
Django==4.2.13
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-jazzmin==3.0.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kombu==5.3.7
packaging==24.0
pipenv==2023.12.1
platformdirs==4.2.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytz==2024.1
redis==5.0.4
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.3
